+++ go/src/problem_1/problem_1_test.go
@@ -6,6 +6,6 @@ import (
 )
 
 func TestProblem1(t *testing.T) {
-	sum := SumOfMultiples([]uint32{3, 5}, 1000)
+	sum := SumOfMultiples([]uint64{3, 5}, 1000)
 	fmt.Printf("Sum of multiples of 3, 5 below 1000 is %v\n", sum)
 }
