+++ go/src/problem_2/problem_2_test.go
@@ -6,7 +6,7 @@ import (
 )
 
 func TestProblem2(t *testing.T) {
-	sum := SumOfFibonacciSeq(func(n uint32) bool {
+	sum := SumOfFibonacciSeq(func(n uint64) bool {
 		return n%2 == 0
 	}, 4000000)
 	fmt.Printf("Sum of even fibinacci sequence under 4 million is %v\n", sum)
